host/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nfcmsr"
version = "0.1.0"
description = "Host CLI for the PN532 + MSR605X pentesting toolkit"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "pn532 toolkit contributors"}]
dependencies = [
    "click>=8.1",
    "pyserial>=3.5",
    "jsonschema>=4.21",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.scripts]
nfcmsr = "nfcmsr.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nfcmsr*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
